summixedlist

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -4,6 +4,6 @@
 from typing import List, Union
 
 
-def sum_mixed_list(mxt_list: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxt_list: List[Union[int, float]]) -> float:
     """Return the sum of a list of floats and integers"""
     return sum(mxt_list)
